Struct rusted_cypher::cypher::result::Row

pub struct Row<'a> {
    // some fields omitted
}

Methods

impl<'a> Row<'a>

fn new(columns: &'a Vec<String>, data: &'a Vec<Value>) -> Self

fn get<T: Deserialize>(&self, column: &str) -> Result<T, GraphError>

Gets the value of a column by its name

The column name must be exactly as it was named in the query. For example, if your query was "MATCH (n:PERSON) RETURN n.name, n.gender", then you should use row.get("n.name") and row.get("n.gender") since this is what the neo4j api produces.

If the column does not exist in the row, an Err is returned with the message "No such column: {column_name}".

fn get_n<T: Deserialize>(&self, column: usize) -> Result<T, GraphError>

Gets the value of a column by order

Column number is 0 based, so the first column is 0, the second is 1 and so on.

If the column number is not within the columns length, and Err is returned with the message "No such column at index {column_number}".