Struct rusted_cypher::cypher::Cypher

pub struct Cypher {
    // some fields omitted
}

Represents the cypher endpoint of a neo4j server

The Cypher struct holds information about the cypher enpoint. It is used to create the queries that are sent to the server.

Methods

impl Cypher

fn new(endpoint: Url, headers: Headers) -> Self

Creates a new Cypher

Its arguments are the cypher transaction endpoint and the HTTP headers containing HTTP Basic Authentication, if needed.

fn query(&self) -> CypherQuery

Creates a new CypherQuery

fn exec<S: Into<Statement>>(&self, statement: S) -> Result<CypherResult, GraphError>

Executes the given Statement

Parameter can be anything that implements Into<Statement>, &str or Statement itself

fn transaction(&self) -> Transaction<Created>

Creates a new Transaction